Changed project structure to and updated class prefixes.
Tapestry can now be imported by dragging the Tapestry.xcodeproj into your project and adding this as to the Header search paths lists in the app project:

"$(BUILT_PRODUCTS_DIR)/include/Headers/"

In Build Phases of the app project add "libTapestry.a" as a library to link with.

In the Other Linker Flags section, add "-ObjC" without quotation marks.